Hw6

See merge request autorobots2018/xrobo!9

+++ core/localization/KMeans.h
@@ -7,13 +7,13 @@
 #include <memory/MemoryCache.h>
 #include <memory/LocalizationBlock.h>
 #include <localization/Logging.h>
-#include <localization/Particle.h>
+#include <common/Particle.h>
 #include <memory/WorldObjectBlock.h>
 
 struct Cluster {
   std::vector<const Particle*> particles;
   Particle centroid;
-  float variance;
+  double variance;
 };
   
 class KMeans {
